[cxx-interop] [test] Clean up some old FRT inheritance tests
These were getting difficult to maintain and are superseded by the
inherit-frt-* tests.

These tests were originally introduced to check for layout
considerations of inheritance, so I've renamed them to reflect that.
