+++ test/Interop/Cxx/foreign-reference/Inputs/module.modulemap
@@ -23,8 +23,8 @@ module Nullable {
   requires cplusplus
 }
 
-module Inheritance {
-  header "inheritance.h"
+module InheritanceLayout {
+  header "inheritance-layout.h"
   requires cplusplus
 }
